refactor(ms2/wrapServerCall): moved to more semantic filename
