+++ src/management-system-v2/components/processes/index.tsx
@@ -45,7 +45,7 @@ import Ability from '@/lib/ability/abilityHelper';
 import ContextMenuArea from './context-menu';
 import { DraggableContext } from './draggable-element';
 import SelectionActions from '../selection-actions';
-import { wrapServerCall } from '@/lib/user-error';
+import { wrapServerCall } from '@/lib/wrap-server-call';
 
 export function canDeleteItems(
   items: ProcessListProcess[],
